Add Leaderboard and theme settings
