+++ web/src/app/leaderboard/layout.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
 
-export default function LegacyLeadpageLayout({ children }: { children: React.ReactNode }) {
+export default function LeaderboardLayout({ children }: { children: React.ReactNode }) {
   return (
     <Suspense fallback={<div className="min-h-screen bg-[var(--nexus-bg)] nexus-bg" aria-hidden />}>
       {children}
