+++ web/src/components/LoginRequiredPanel.tsx
@@ -25,7 +25,7 @@ export function LoginRequiredPanel({
           {cta}
         </Link>
         <Link
-          href="/leadpage"
+          href="/leaderboard"
           className="rounded-xl border border-[rgba(138,149,166,0.18)] bg-[rgba(6,8,11,0.35)] px-3 py-2 text-[11px] text-[rgba(226,232,240,0.88)] hover:border-[rgba(0,212,170,0.32)] hover:text-white"
         >
           Back to Leaderboard
