+++ web/src/components/NexusConsoleHeader.tsx
@@ -127,7 +127,7 @@ export function NexusConsoleHeader({
         </div>
 
         <div className="w-full mt-2 border-t border-[var(--nexus-rule-soft)] pt-2 flex flex-wrap items-center justify-start gap-3">
-          <Suspense fallback={<div className="h-10 w-full max-w-md rounded-lg bg-[rgba(6,8,11,0.35)]" />}>
+          <Suspense fallback={<div className="h-10 w-full max-w-md rounded-lg bg-[var(--nexus-surface)]" />}>
             <NexusHeaderNav
               active="nexus"
               variant="console"
@@ -139,22 +139,22 @@ export function NexusConsoleHeader({
           <span
             className={`rounded-lg border px-2 py-1 text-[10px] font-mono ${
               wsConnected
-                ? "border-[rgba(0,212,170,0.22)] bg-[rgba(0,212,170,0.08)] text-[rgba(226,232,240,0.92)]"
-                : "border-[rgba(138,149,166,0.22)] bg-[rgba(138,149,166,0.10)] text-[rgba(226,232,240,0.86)]"
+                ? "border-[rgba(0,212,170,0.22)] bg-[rgba(0,212,170,0.08)] text-[var(--nexus-text)]"
+                : "border-[var(--nexus-border)] bg-[var(--nexus-surface)] text-[var(--nexus-muted)]"
             }`}
             title="WebSocket stream connection"
           >
             {wsConnected ? "stream: connected" : "stream: offline"}
           </span>
           <span
-            className="rounded-lg border border-[rgba(138,149,166,0.18)] bg-[rgba(0,0,0,0.15)] px-2 py-1 text-[10px] font-mono text-[var(--nexus-muted)]"
+            className="rounded-lg border border-[var(--nexus-border)] bg-[var(--nexus-surface)] px-2 py-1 text-[10px] font-mono text-[var(--nexus-muted)]"
             title="Last payload update time"
           >
             {loading ? "updating…" : lastUpdateIso ? `last: ${new Date(lastUpdateIso).toLocaleTimeString()}` : "last: —"}
           </span>
           {metadata?.run_id ? (
-            <span className="rounded-lg border border-[rgba(138,149,166,0.18)] bg-[rgba(0,0,0,0.15)] px-2 py-1 text-[10px] font-mono text-[var(--nexus-muted)]">
-              run: <span className="text-[rgba(226,232,240,0.92)]">{metadata.run_id}</span>
+            <span className="rounded-lg border border-[var(--nexus-border)] bg-[var(--nexus-surface)] px-2 py-1 text-[10px] font-mono text-[var(--nexus-muted)]">
+              run: <span className="text-[var(--nexus-text)]">{metadata.run_id}</span>
             </span>
           ) : null}
           <KpiStrip kpis={metadata?.kpis ?? {}} />
