+++ web/src/components/NexusSubtabs.tsx
@@ -31,7 +31,7 @@ function Tab({
 
 function isPathActive(pathname: string, href: string): boolean {
   if (href === "/") return pathname === "/";
-  if (href === "/leadpage") return pathname === "/leadpage" || pathname.startsWith("/leadpage/");
+  if (href === "/leaderboard") return pathname === "/leaderboard" || pathname.startsWith("/leaderboard/");
   if (href === "/console") return pathname === "/console";
   if (href.startsWith("/console?")) return pathname === "/console";
   return pathname === href || pathname.startsWith(`${href}/`);
@@ -60,7 +60,7 @@ export function NexusSubtabs({ active }: { active: HeaderNavMode }) {
   const tabs = useMemo(() => {
     if (active === "observe") {
       return [
-        { href: "/leadpage", label: "Results", title: "Leaderboard results" },
+        { href: "/leaderboard", label: "Results", title: "Leaderboard results" },
         { href: "/feed", label: "Signals", title: "Provider signals feed" },
       ];
     }
